Refactor SignalQueue path in accordance with recent organization changes
