+++ src/ECS/SignalQueue/DataCargo/MouseEvents/MouseMoveSignalData.cpp
@@ -3,7 +3,7 @@
 //
 
 #include "MouseMoveSignalData.h"
-#include "ECS/SignalSystem/Signal.h"
+#include "ECS/SignalQueue/Signal.h"
 #include "ECS/Util.h"
 
 using namespace ztgk;
